+++ tests/flow/test_config.py
@@ -326,3 +326,90 @@ class testConfig(FlowTestsBase):
         creation_buffer_size = self.db.config_get("NODE_CREATION_BUFFER")
         expected_response = 1024
         self.env.assertEqual(creation_buffer_size, expected_response)
+
+import stat
+import shutil
+import tempfile
+
+class testConfigTempFolder:
+    def teardown_method(self):
+        if hasattr(self, 'conn'):
+            self.conn.shutdown()
+
+    def set_temp_folder(self, path):
+        module_args = f"TEMP_FOLDER {path}"
+        self.env, self.db = Env(moduleArgs=module_args, enableDebugCommand=True)
+
+        if SANITIZER:
+            self.env.skip()
+
+        self.conn = self.env.getConnection()
+
+    def test_01_temp_folder_is_file(self):
+        # try setting TEMP_FOLDER to a file
+        # expecting config update to fail
+        fd, file_path = tempfile.mkstemp()
+        os.close(fd)
+
+        # try updating TEMP_FOLDER
+        try:
+            self.set_temp_folder(file_path)
+            # setting TEMP_FOLDER to a file should have failed
+            self.env.assertFalse(True)
+        except Exception:
+            pass
+
+    def test_02_temp_folder_not_exist(self):
+        # try setting TEMP_FOLDER to a non existing folder
+        # expecting config update to fail
+
+        # make sure path doesn't exists
+        non_existent = "/tmp/falkordb_nonexistent_dir"
+        if os.path.exists(non_existent):
+            shutil.rmtree(non_existent)
+
+        # try updating TEMP_FOLDER
+        try:
+            self.set_temp_folder(non_existent)
+            self.env.assertFalse(True)
+        except Exception:
+            pass
+
+    def test_03_temp_folder_no_permission(self):
+        # try setting TEMP_FOLDER to a folder which we can't write to
+        # expecting config update to fail, as write access is mandatory
+
+        # create a temp folder with no write access
+        no_perm_dir = tempfile.mkdtemp()
+        os.chmod(no_perm_dir, stat.S_IREAD)
+
+        # check if directory is truly unwritable
+        if os.access(no_perm_dir, os.W_OK):
+            env, _ = Env(enableDebugCommand=True)
+            env.skip()
+
+        # try updating TEMP_FOLDER
+        try:
+            self.set_temp_folder(no_perm_dir)
+            # setting TEMP_FOLDER to a non writeable folder should have failed
+            self.env.assertFalse(True)
+        except Exception:
+            pass
+        finally:
+            # clean up
+            os.chmod(no_perm_dir, stat.S_IWUSR | stat.S_IREAD | stat.S_IXUSR)
+            shutil.rmtree(no_perm_dir)
+
+    def test_04_temp_folder_exists_success(self):
+        # try setting TEMP_FOLDER to a valid folder
+        # expecting config update to succeed
+
+        valid_dir = tempfile.mkdtemp()
+
+        try:
+            self.set_temp_folder(valid_dir)
+            self.env.assertEqual(self.db.config_get("TEMP_FOLDER"), valid_dir)
+        finally:
+            # clean up
+            shutil.rmtree(valid_dir)
+
