--- tests/flow/test_graph_copy_temp_folder.py
@@ -1,67 +1,0 @@
-from common import Env, SANITIZER
-import os
-import tempfile
-import shutil
-import stat
-import pytest
-
-
-class testGraphCopyTempFolder:
-    def teardown_method(self):
-        # Only attempt delete if setup completed and attributes exist
-        if hasattr(self, 'conn') and hasattr(self, 'src') and hasattr(self, 'dest') and self.conn:
-            try:
-                self.conn.delete(self.src, self.dest)
-            except Exception:
-                pass
-
-    def set_temp_folder(self, path):
-        module_args = f"TEMP_FOLDER {path}"
-        self.env, self.db = Env(moduleArgs=module_args, enableDebugCommand=True)
-        if SANITIZER:
-            self.env.skip()
-        self.conn = self.env.getConnection()
-        self.src = 'temp_src_graph'
-        self.dest = 'temp_dest_graph'
-        src_graph = self.db.select_graph(self.src)
-        src_graph.query("RETURN 1")
-
-    def test_01_temp_folder_is_file(self):
-        fd, file_path = tempfile.mkstemp()
-        os.close(fd)
-        try:
-            with pytest.raises(Exception):
-                self.set_temp_folder(file_path)
-        finally:
-            os.remove(file_path)
-
-    def test_02_temp_folder_not_exist(self):
-        non_existent = "/tmp/falkordb_nonexistent_dir"
-        if os.path.exists(non_existent):
-            shutil.rmtree(non_existent)
-        with pytest.raises(Exception):
-            self.set_temp_folder(non_existent)
-
-    def test_03_temp_folder_no_permission(self):
-        no_perm_dir = tempfile.mkdtemp()
-        os.chmod(no_perm_dir, stat.S_IREAD)
-        # Check if directory is truly unwritable
-        if os.access(no_perm_dir, os.W_OK):
-            env, _ = Env(enableDebugCommand=True)
-            env.skip()
-        try:
-            with pytest.raises(Exception):
-                self.set_temp_folder(no_perm_dir)
-        finally:
-            os.chmod(no_perm_dir, stat.S_IWUSR | stat.S_IREAD | stat.S_IXUSR)
-            shutil.rmtree(no_perm_dir)
-
-    def test_04_temp_folder_exists_success(self):
-        valid_dir = tempfile.mkdtemp()
-        self.set_temp_folder(valid_dir)
-        try:
-            self.conn.execute_command("GRAPH.COPY", self.src, self.dest)
-            # Should succeed, check dest exists
-            assert self.conn.type(self.dest) == 'graphdata'
-        finally:
-            shutil.rmtree(valid_dir)
